+++ src/wave/heroes/donald_rebeli/mod.rs
@@ -1,0 +1,2 @@
+use super::test_hero_string;
+test_hero_string!("Donald Rebeli");