+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Hero Lydia (w/o Jolly Water fight)
+- Hero Donald Rebeli (w/o Aristocratic Duel)
 
 ## [0.0.17] - 2023-10-28